test/vhost: remove tests for live migration between different hosts
The test case 3 a/b were tests for live migration were to be executed
between different hosts.

Since most of the tests in CI are executed within single host,
those were to be delegated to separate job that would use more hosts.

Yet for some time this went untested and state of the scripts is unclear.
Let's remove it.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I6bab2c0b85da87d29cfccabff2103812b181f246
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/5037
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Paul Luse <paul.e.luse@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>